# strict-server/internal/api/gen/types.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.1.0 DO NOT EDIT.
package gen

// ErrorResult defines model for ErrorResult.
type ErrorResult struct {
	Code       string `json:"code"`
	Message    string `json:"message"`
	StatusCode int    `json:"statusCode"`
}

// User defines model for User.
type User struct {
	FirstName string `json:"firstName" validate:"required"`
	LastName  string `json:"lastName" validate:"required"`
	UserEmail string `json:"userEmail" validate:"email"`
}

// UserWithId defines model for UserWithId.
type UserWithId struct {
	FirstName string `json:"firstName" validate:"required"`
	Id        uint32 `json:"id" validate:"gte=0"`
	LastName  string `json:"lastName" validate:"required"`
	UserEmail string `json:"userEmail" validate:"email"`
}

// CreateUserJSONRequestBody defines body for CreateUser for application/json ContentType.
type CreateUserJSONRequestBody = User

// UpdateUserJSONRequestBody defines body for UpdateUser for application/json ContentType.
type UpdateUserJSONRequestBody = User