# strict-server/strict-server-users-openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: API-First User Service
  title: User-Service
  version: '1.0'
servers:
  - url: 'http://localhost:8080/v1'
tags:
  - name: users
paths:
  '/users':
    post:
      description: Create User
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithId'
          description: User created
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
          description: Unprocessable Content
      summary: Create a user object
      tags:
        - users
  '/users/{id}':
    put:
      description: Update a User
      operationId: updateUser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: uint32
            example: 1
            x-oapi-codegen-extra-tags:
              validate: gte=0
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithId'
          description: User accepted
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
          description: Bad Request
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
          description: Unprocessable Content
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
          description: Internal Server Error
      summary: Update a user object by id
      tags:
        - users
    get:
      description: Get a User by id
      operationId: getUser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: uint32
            example: 1
            x-oapi-codegen-extra-tags:
              validate: gte=0
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithId'
          description: User retrieved
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
          description: Bad Request
      summary: Get a user object by id
      tags:
        - users
    delete:
      description: Delete a User by id
      operationId: deleteUser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: uint32
            example: 1
            x-oapi-codegen-extra-tags:
              validate: gte=0
      responses:
        '204':
          description: User deleted
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
          description: Bad Request
      summary: Delete a user object by id
      tags:
        - users
components:
  parameters: {}
  responses: {}
  schemas:
    User:
      type: object
      required:
        - firstName
        - lastName
        - userEmail
      properties:
        firstName:
          type: string
          example: John
          x-oapi-codegen-extra-tags:
            validate: required
        lastName:
          type: string
          example: Doe
          x-oapi-codegen-extra-tags:
            validate: required
        userEmail:
          type: string
          example: john.doe@mymail.com
          x-oapi-codegen-extra-tags:
            validate: email
    UserWithId:
      allOf:
        - type: object
          required:
            - id
          properties:
            id:
              type: integer
              format: uint32
              example: 1
              x-oapi-codegen-extra-tags:
                validate: gte=0
        - $ref: '#/components/schemas/User'
    ErrorResult:
      properties:
        statusCode:
          type: integer
        code:
          type: string
        message:
          type: string
      required:
        - statusCode
        - code
        - message
      title: ErrorResult
      type: object